requirement.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
requests
ipykernel
bs4
pypdf
faiss-cpu
wikipedia
arxiv